--- src/features/Product/PostProduct/store/PostProudctSlice.js
@@ -1,15 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-const initialState = {
-  productName: '',
-  category: '',
-  description: '',
-  tags: '',
-  price: '',
-  images: [],
-};
-
-const productSlice = createSlice({
-  name: 'product',
-  initialState,
-});
